+++ deploy/complete/terraform/policies.tf
@@ -3,17 +3,17 @@
 # 
 
 resource "oci_identity_dynamic_group" "oke_nodes_dg" {
-  name           = "${lower(trimspace(var.app_name))}-oke-cluster-dg-${random_string.deploy_id.result}"
+  name           = "${local.app_name_normalized}-oke-cluster-dg-${random_string.deploy_id.result}"
   description    = "${var.app_name} Cluster Dynamic Group"
   compartment_id = var.tenancy_ocid
-  matching_rule  = "ALL {instance.compartment.id = '${local.oke_compartment_ocid}'}"
+  matching_rule  = "ANY {ALL {instance.compartment.id = '${local.oke_compartment_ocid}'},ALL {resource.type = 'cluster', resource.compartment.id = '${local.oke_compartment_ocid}'}}"
 
   provider = oci.home_region
 
   count = var.create_dynamic_group_for_nodes_in_compartment ? 1 : 0
 }
 resource "oci_identity_policy" "oke_compartment_policies" {
-  name           = "${lower(trimspace(var.app_name))}-oke-cluster-compartment-policies-${random_string.deploy_id.result}"
+  name           = "${local.app_name_normalized}-oke-cluster-compartment-policies-${random_string.deploy_id.result}"
   description    = "${var.app_name} OKE Cluster Compartment Policies"
   compartment_id = local.oke_compartment_ocid
   statements     = local.oke_compartment_statements
@@ -25,7 +25,7 @@ resource "oci_identity_policy" "oke_compartment_policies" {
   count = var.create_compartment_policies ? 1 : 0
 }
 resource "oci_identity_policy" "kms_compartment_policies" {
-  name           = "${lower(trimspace(var.app_name))}-kms-compartment-policies-${random_string.deploy_id.result}"
+  name           = "${local.app_name_normalized}-kms-compartment-policies-${random_string.deploy_id.result}"
   description    = "${var.app_name} KMS Compartment Policies"
   compartment_id = local.oke_compartment_ocid
   statements     = local.kms_compartment_statements
@@ -38,7 +38,7 @@ resource "oci_identity_policy" "kms_compartment_policies" {
 }
 
 resource "oci_identity_policy" "oke_tenancy_policies" {
-  name           = "${lower(trimspace(var.app_name))}-oke-cluster-tenancy-policies-${random_string.deploy_id.result}"
+  name           = "${local.app_name_normalized}-oke-cluster-tenancy-policies-${random_string.deploy_id.result}"
   description    = "${var.app_name} OKE Cluster Tenancy Policies"
   compartment_id = var.tenancy_ocid
   statements     = local.oke_tenancy_statements
@@ -56,7 +56,8 @@ locals {
   )
   oke_compartment_statements = concat(
     local.oci_grafana_logs_statements,
-    var.use_encryption_from_oci_vault ? local.allow_oke_use_oci_vault_keys_statements : []
+    var.use_encryption_from_oci_vault ? local.allow_oke_use_oci_vault_keys_statements : [],
+    var.cluster_autoscaler_enabled ? local.cluster_autoscaler_statements : []
   )
   kms_compartment_statements = concat(
     local.allow_group_manage_vault_keys_statements